requirements.txt
# --- Core Web Framework ---
fastapi>=0.100.0
uvicorn[standard]>=0.22.0

# --- Database ---
SQLAlchemy>=2.0.0
alembic>=1.11.0
# For SQLite (local dev), no psycopg needed. Remove if unused.
psycopg[binary]>=3.1.0; extra == 'pg'

# --- Pydantic, Validation & Parsing ---
pydantic>=2.0.0
email-validator>=1.3.0
python-multipart>=0.0.6

# --- Authentication ---
passlib[bcrypt]>=1.7.0
argon2-cffi>=23.1.0
python-jose[cryptography]>=3.3.0

# --- gRPC ---
grpcio>=1.64.0
grpcio-tools>=1.64.0
